refactor(core): extract fresh-install schema migration from ledger.ts
Move the fresh-install schema construction and inline column/table
migrations out of Ledger.migrate() into a dedicated
packages/core/src/ledger-schema.ts module (migrateLedgerSchema),
operating on a raw bun:sqlite Database so it can be exercised
independently of the rest of the Ledger class. Ledger.migrate() now
delegates to it; the exported Ledger surface and every call site are
unchanged.

addColumnIfMissing moved with it and is now exported for direct
testing (ledger-extra.test.ts's identifier-guard test previously
reached into Ledger's private method via an unsafe cast; it now
calls the exported function against a throwaway in-memory database).

Added a deterministic schema snapshot test (ledger.test.ts) that
dumps sqlite_master (every table/index/trigger's CREATE statement)
plus schema_version for a fresh install and pins it via
toMatchSnapshot -- proof the extraction produced a byte-identical
schema. Also verified manually pre/post-extraction with an ad hoc
dump script (diff was empty).

ledger.ts: 3929 -> 3366 lines. New ledger-schema.ts: 590 lines.

Records this as an in-progress step (not the full split) on the
ledger-split ROADMAP item.

Fixes #452
