+++ packages/core/__tests__/ledger.test.ts
@@ -70,6 +70,27 @@ describe("Ledger schema migration", () => {
     });
     expect(ledger.listPendingQueueAfterId(before).map((row) => row.dedupe_key)).toEqual(["a"]);
   });
+
+  it("fresh-install schema (tables, columns, indexes, triggers, schema_version) matches the snapshot", () => {
+    // Deterministic proof that extracting the migration DDL into
+    // ledger-schema.ts (issue #452) didn't change what a fresh install
+    // produces. sqlite_master carries every table/index/trigger's exact
+    // CREATE statement; schema_version is the migration's own version
+    // marker. If this snapshot ever needs updating, confirm the DDL change
+    // was intentional before accepting the new baseline.
+    const db = (ledger as unknown as { db: Database }).db;
+    const objects = db
+      .query(
+        `SELECT type, name, tbl_name, sql FROM sqlite_master
+         WHERE sql IS NOT NULL AND name NOT LIKE 'sqlite_%'
+         ORDER BY type, name`,
+      )
+      .all() as Array<{ type: string; name: string; tbl_name: string; sql: string }>;
+    const version = db.query("SELECT version FROM schema_version").all() as Array<{
+      version: number;
+    }>;
+    expect({ objects, version }).toMatchSnapshot();
+  });
 });
 
 describe("Ledger receipts + prospects + spend rollups", () => {
